+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # CHANGELOG
 
+## v0.0.83 [unreleased]
+
+- `ryo3-jiff`
+  - more timestamp testing
+
+---
+
 ## v0.0.82 [2026-01-27]
 
 - `ryo3-http`